+++ Frontend/admin-portal/src/firebase-config.js
@@ -6,13 +6,12 @@ import { getFirestore } from "firebase/firestore";
 import { getStorage } from "firebase/storage";
 
 const firebaseConfig = {
-    apiKey: "AIzaSyAptP2cM_xj764rrwC4FRnbmnQJwFsLvFM",
-    authDomain: "p3-auth-762da.firebaseapp.com",
-    projectId: "p3-auth-762da",
-    storageBucket: "p3-auth-762da.firebasestorage.app",
-    messagingSenderId: "1079324928317",
-    appId: "1:1079324928317:web:b554f43a8a42ea3bbe3d54",
-    measurementId: "G-8KG5C0D6EL"
+    apiKey: "AIzaSyCzckS2MVjMby8dvKDE8wtIWzM2XewvnUg",
+    authDomain: "dismissal-cloud.firebaseapp.com",
+    projectId: "dismissal-cloud",
+    storageBucket: "dismissal-cloud.firebasestorage.app",
+    messagingSenderId: "177955649483",
+    appId: "1:177955649483:web:fc42c84772055e9989284e"
   };
 
 const app = initializeApp(firebaseConfig);